requirements.txt
python-dotenv
audioop-lts
discord
pillow